+++ app/Item.php
@@ -312,9 +312,15 @@ class Item extends Model implements IndexableModel, TranslatableContract
         $authors_array = $this->makeArray($this->author);
         $authors = array();
         foreach ($authors_array as $author) {
-            $authors[$author] = preg_replace('/^([^,]*),\s*(.*)$/', '$2 $1', $author);
+            $authors[$author] = formatName($author);
         }
+        
         return $authors;
+    }
+
+    public function getAuthorsFormattedAttribute($value)
+    {
+        return array_map( function($a){return formatName($a);}, $this->authors) ;
     }
 
     public function getFirstAuthorAttribute($value)
@@ -510,7 +516,7 @@ class Item extends Model implements IndexableModel, TranslatableContract
         $used_authorities = array();
         $authorities_with_link = array();
         $not_authorities_with_link = array();
-        foreach ($this->authorities as $authority) {
+        foreach ($this->authorities->sortBy('name') as $authority) {
             if ($authority->pivot->role != 'autor/author') {
                 $not_authorities_with_link[] = '<a class="underline" href="'. $authority->getUrl() .'">'. $authority->formated_name .'</a>'
                     .' &ndash; ' . Authority::formatMultiAttribute($authority->pivot->role);